pkg/api pagination fixes
CurrentPage() is expected to return the page currently retrieved, which
page the data in the passed array belongs to. Update test to reflect
that.

currentPage was incremented before doing the API request, meaning it
returned the page last tried to be retrieved instead of the last one
retrieved successfully. Now currentPage+1 is requested and currentPage
is incremented on success.

Add check for query argument page != 0 in pagination test.