✨💥: fundamentally change List via channel
Previous to this commit, we let users create a channel variable and pass
a pointer to this to the AsObjectChannel option, which initialized it
and made the API client spawn a goroutine to fill it.

We now let the user create a "ObjectChannelCloser" via NewObjectChannel
and pass that to the (renamed) ObjectChannel option. Via this new
ObjectChannelCloser, the user can retrieve the go channel to stream
objects as previous, but also cancel the listing - which was not
possible before and leaked goroutines whenever the channel was not read
until it finished.