✅🩹: allow unknown fields in most responses
We configured the JSON decoder in use by the generic client to
`DisallowUnknownFields` before, because encoding/json does not know
about required fields, resulting in every json object being decodable
into every Go struct, but we want to make sure the response is actually
the type we expect.

Turns out this is only really needed for checking which style of
pagination response we got, not for objects and we already have
different code paths for decoding pagination responses and objects. This
commit allows unknown fields for objects.

We previously had a test case for a valid pagination response with
invalid objects in it. We did not see that in real life (as far as we
remember), so we removed that test case - hoping the Engine does not
throw such responses at us.