💩 Location: complicated magic for float coords
I don't want to loose it, as the logic is quite nice and might be useful
in the future somewhere, but definitely not useful here - especially as
we defined "client" as thin layer above the API and everything below
pkg/apis/ to be a "client" today.

Will change the coords to *string in next commit, removing all this
complicated code.