Makefile: keep going through all tests
This commit changes the behavior of our test suites from "abort once a
suite has failed" to "run through all suites while report failures".

This helps us continue working and relying on the CI system when there
is a regression in an API as we can just ignore failures in a specific
API when everything else is still green.