💡: clarify we normally don't want full E2E tests
LBaaS v1 bindings include a full E2E test checking if we can actually
reach a service via the given test configuration. This check should not
live in go-anxcloud but rather in LBaaS project, but here we are and we
rather have this test here than nowhere..

This commit extracts those parts into a new file and adds a comment to
clarify we do not want such full E2E tests for just everything.