Add missing anxcloud:identifier struct tags
When working on our ake-controller-manager together with the usage of
the go-anxcloud mock module, I've stumbled across this issue for the
common.PartialResource.

This error is only occuring because the [makeObjectIdentifiable] function
requires the `anxcloud:"identifier"` tag to be set. I haven't touched the
code in the function itself, as any panic or error caused by it might
break existing callers or usecases I haven't considered.

[makeObjectIdentifiable]: pkg/api/mock/utils.go