Provide generic IPAM API implementation
This is the first iteration, providing most of the old address and
prefix API. However, most of the important stuff is implemented.
Unfortunately, it took me waaay to long to get to this state, mainly due
to timeouts in the Engine and missing knowledge of the client internals.

Although the tests can run mostly parallel, we might switch to a serial
execution to ensure that the API isn't bombarded with endless requests.

Co-authored-by: Mara Sophie Grosch <mgrosch@anexia.com>
Ticket: SYSENG-923