Add dedicated RateLimitError for HTTP code 429 (#438)
Rate limits got introduced by the Anexia Engine at some point in the
second half of 2024. Most of our integrations were not prepared for
this, especially CCM is suffering from this.

In order to detect rate limits sent by the Engine, a new RateLimitError
was added, together with a timestamp after which the requests should be
retried again.

I decided to introduce a complex type with a placeholder value for this
timestamp right now. This allows us to build other integrations on top
of it and when we later update the timestamp with the correct value of
the Retry-After header (ENGSUP-9027), all existing implementations would
benefit from that without any further manual changes.

Related: ANXKUBE-1281